backend/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
dtale==3.18.2
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Compress==1.18
fonttools==4.59.0
future==1.0.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kaleido==1.0.0
kiwisolver==1.4.8
logistro==1.1.0
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.5
missingno==0.5.2
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
orjson==3.11.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
pkginfo==1.12.1.2
plotly==6.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzstd==0.17.0
requests==2.32.4
retrying==1.4.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sklearn-compat==0.1.3
soupsieve==2.7
squarify==0.4.4
statsmodels==0.14.5
strsimpy==0.2.1
sweetviz==2.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.0.6
xarray==2025.7.1
xgboost==3.0.3
xlrd==2.0.2
zipp==3.23.0
python-dotenv==1.1.1
